backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic-settings==2.7.1
motor==3.6.0
beanie==1.27.0
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.159.0
openpyxl==3.1.5
pandas==2.2.3
structlog==24.4.0
python-multipart==0.0.20
starlette==0.41.3
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1
itsdangerous==2.2.0
mongomock-motor==0.0.34
